+++ playground/src/utils/index.ts
@@ -15,6 +15,7 @@ export { isSensitiveEnvVar } from "./env-vars";
 export { renderFullLogMessage } from "./render-log-message";
 export { truncateWithEllipsis } from "./truncate";
 export { parseEmbeddedConfig } from "./config-parser";
+export { getHttpMethodTextColor } from "./http-method-color";
 export { safeParseJson } from "./safe-parse-json";
 export function formatDate(d: Date | string) {
   return format(new Date(d), "HH:mm:ss.SSS");
@@ -196,19 +197,6 @@ export function formatHeaders(headers: Record<string, string>): string {
   return Object.entries(headers)
     .map(([key, value]) => `${key}: ${value}`)
     .join("\n");
-}
-
-export function getHttpMethodTextColor(method: string) {
-  return {
-    GET: "text-info",
-    POST: "text-success",
-    PUT: "text-warning",
-    PATCH: "text-warning",
-    DELETE: "text-danger",
-    OPTIONS: "text-info",
-    HEAD: "text-info",
-    WS: "text-success",
-  }[String(method).toUpperCase()];
 }
 
 export function constructPlaygroundBody(bodyValue: string): PlaygroundBody {
